requirements.txt
py-cord~=2.6
python-dotenv